+++ changelog.md
@@ -43,6 +43,7 @@
 - Added class `Projects/Secrets/DefaultBundle`
 - Added class `Projects/Tags`
 - Added class `Projects/Vault`
+- Added class `Projects/Topics`
 - **[BREAKING]** Renamed method and class `GraphExecution` to `GraphExecutions`
 - **[BREAKING]** Renamed method and class `Job` to `Jobs`
 - **[BREAKING]** Renamed method and class `Repository` to `Repositories`